MODLD-1069 - refactor tests to extend PostAndPutResourceIT for improved PUT request handling
