+++ src/test/java/org/folio/linked/data/e2e/mappings/instance/bookformat/BookFormatIT.java
@@ -7,7 +7,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import lombok.SneakyThrows;
 import org.folio.linked.data.e2e.base.IntegrationTest;
-import org.folio.linked.data.e2e.mappings.PostResourceIT;
+import org.folio.linked.data.e2e.mappings.PostAndPutResourceIT;
 import org.folio.linked.data.model.entity.Resource;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.test.context.ActiveProfiles;
@@ -15,7 +15,7 @@ import org.springframework.test.web.servlet.ResultActions;
 
 @IntegrationTest
 @ActiveProfiles({STANDALONE_PROFILE, STANDALONE_TEST_PROFILE})
-class BookFormatIT extends PostResourceIT {
+class BookFormatIT extends PostAndPutResourceIT {
 
   @BeforeEach
   void createWork() {
